Add Game 1 scene and move services to the game installer
