+++ Assets/PolygonNature/Shaders/LWRP.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d83f241364eceaf4081bf2b3ee6ee87b
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
