+++ Assets/PolygonNature/URP_Extract_Me.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a064949e8b7c31e4699684de4fb7ff30
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
