+++ Assets/Scripts/Services/Interfaces/IInteractionService.cs
@@ -4,7 +4,6 @@ namespace Services.Interfaces
 {
     public interface IInteractionService
     {
-        void Configure(float interactionDistance, LayerMask interactionLayer);
         void SetCamera(Camera mainCamera);
         void CanGetItems(bool canGetItems);
     }
