+++ ProjectSettings/EditorBuildSettings.asset
@@ -11,6 +11,9 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/MainMenu.unity
     guid: 7350fb21b955bfa4f95476da6ef466be
+  - enabled: 1
+    path: Assets/Scenes/Game 1.unity
+    guid: 52f768539bda50049add359c7131a53f
   m_configObjects:
     com.unity.addressableassets: {fileID: 11400000, guid: 111ce44e8890acd4fbecbc32bdf5c21f,
       type: 2}
